Apply generic msg
