+++ packages/desmos-std/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod subspaces;
 pub mod types;
+pub mod msg;
