+++ packages/desmos-std/src/subspaces/mod.rs
@@ -2,8 +2,8 @@
 pub mod mocks;
 
 pub mod models;
-pub mod msg_router;
-pub mod msg_routes;
+pub mod msg_builder;
+pub mod msg;
 pub mod querier;
 pub mod query_router;
 pub mod query_types;
